+++ test/src/test_program.cpp
@@ -19,16 +19,17 @@ TEST(test_program, void_command)
 {
     std::istringstream test_in;
     std::ostringstream test_out;
-    skipper::program p("dummy help text", test_in, test_out);
+    skipper::program p("dummy help text",
+                       skipper::in = test_in,
+                       skipper::out = test_out,
+                       skipper::print_help_on_run = false,
+                       skipper::ready_indicator = "");
 
     stub::call<void(void)> m_print_function;
     auto print_function = [&](){m_print_function();};
 
     p.add_command("p", "print something", print_function);
     test_in.str("p\n p\n p\n q\n");
-
-    p.set_print_help_on_run(false);
-    p.set_ready_indicator("");
     p.run();
     EXPECT_EQ(3U, m_print_function.calls());
     EXPECT_EQ("", test_out.str());
@@ -38,16 +39,17 @@ TEST(test_program, int_set_command)
 {
     std::istringstream test_in;
     std::ostringstream test_out;
-    skipper::program p("dummy help text", test_in, test_out);
+    skipper::program p("dummy help text",
+                       skipper::in = test_in,
+                       skipper::out = test_out,
+                       skipper::print_help_on_run = false,
+                       skipper::ready_indicator = "");
 
     stub::call<void(int)> m_function;
     std::function<void(int)> function = [&](int value){m_function(value);};
     p.add_command<int>("a", "help", function, skipper::set<int>({0,7,42,1}));
 
     test_in.str("a\n 42\n a\n -10\n a\n 0\n a\n 7 a\n 11\n");
-
-    p.set_print_help_on_run(false);
-    p.set_ready_indicator("");
     p.run();
     EXPECT_EQ(3U, m_function.calls());
     EXPECT_TRUE(m_function.expect_calls().with(42).with(0).with(7).to_bool());
@@ -57,7 +59,11 @@ TEST(test_program, float_range_command)
 {
     std::istringstream test_in;
     std::ostringstream test_out;
-    skipper::program p("dummy help text", test_in, test_out);
+    skipper::program p("dummy help text",
+                       skipper::in = test_in,
+                       skipper::out = test_out,
+                       skipper::print_help_on_run = false,
+                       skipper::ready_indicator = "");
 
     stub::call<void(float)> m_function;
     std::function<void(float)> function = [&](float value){m_function(value);};
@@ -65,9 +71,6 @@ TEST(test_program, float_range_command)
         skipper::range<float>(-4.0F, 7.2F));
 
     test_in.str("a\n 42\n a\n -1\n a\n 0\n a\n 7\n a\n 11\n");
-
-    p.set_print_help_on_run(false);
-    p.set_ready_indicator("");
     p.run();
     EXPECT_EQ(3U, m_function.calls());
     EXPECT_TRUE(m_function.expect_calls().with(-1).with(0).with(7).to_bool());
@@ -77,16 +80,17 @@ TEST(test_program, trigger_errors)
 {
     std::istringstream test_in;
     std::ostringstream test_out;
-    skipper::program p("dummy help text", test_in, test_out);
+    skipper::program p("dummy help text",
+                       skipper::in = test_in,
+                       skipper::out = test_out,
+                       skipper::print_help_on_run = false,
+                       skipper::ready_indicator = "");
 
     stub::call<void(int)> m_function;
     std::function<void(int)> function = [&](int value){m_function(value);};
     p.add_command<int>("a", "help", function, skipper::range<int>(0, 10));
 
     test_in.str("wrong key\n a\n wrong input\n q\n");
-
-    p.set_print_help_on_run(false);
-    p.set_ready_indicator("");
     p.run();
     EXPECT_NE("", test_out.str());
 }